test: move `permissions.test.ts` file
